multi: expose git info in long_version of CLI config
