+++ lib/build.rs
@@ -1,9 +1,23 @@
 use std::{
     env, fs,
     path::{Path, PathBuf},
+    process::Command,
 };
 
 use prost::Message;
+
+fn get_git_hash() -> Result<String, Box<dyn std::error::Error>> {
+    let args = ["describe", "--always", "--dirty"];
+    let output = Command::new("git").args(args).output()?;
+
+    if !output.status.success() {
+        return Err(format!("Failed to execute `git {}`", args.join(" ")).into());
+    }
+
+    let hash = String::from_utf8(output.stdout)?.trim().to_string();
+
+    Ok(hash)
+}
 
 fn compile_protos_with_config<F>(
     file_descriptor_path: impl AsRef<Path>,
@@ -25,6 +39,19 @@ where
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
+    // Set git hash as environment variable for runtime access
+    match get_git_hash() {
+        Ok(hash) => {
+            println!("cargo:rustc-env=GIT_HASH={}", hash);
+            println!("cargo:rerun-if-changed=.git/HEAD");
+            println!("cargo:rerun-if-changed=.git/refs/heads");
+        }
+        Err(e) => {
+            println!("cargo:warning=Failed to get git hash: {}", e);
+            println!("cargo:rustc-env=GIT_HASH=unknown");
+        }
+    }
+
     const COMMON_PROTO: &str = "../cusf_sidechain_proto/proto/cusf/common/v1/common.proto";
     const CRYPTO_PROTO: &str = "../cusf_sidechain_proto/proto/cusf/crypto/v1/crypto.proto";
     const MAINCHAIN_COMMON_PROTO: &str =
